@@ -1,85 +1,117 @@
+# name: personal_expenses_2
+# description: A new Flutter project.
+
+# # The following line prevents the package from being accidentally published to
+# # pub.dev using `pub publish`. This is preferred for private packages.
+# publish_to: "none" # Remove this line if you wish to publish to pub.dev
+
+# # The following defines the version and build number for your application.
+# # A version number is three numbers separated by dots, like 1.2.43
+# # followed by an optional build number separated by a +.
+# # Both the version and the builder number may be overridden in flutter
+# # build by specifying --build-name and --build-number, respectively.
+# # In Android, build-name is used as versionName while build-number used as versionCode.
+# # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# # Read more about iOS versioning at
+# # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+# version: 1.0.0+1
+
+# environment:
+#   # sdk: ">=2.7.0 <3.0.0"
+#   sdk: ">=2.12.0 <4.0.0"
+
+# dependencies:
+#   flutter:
+#     sdk: flutter
+
+#   # The following adds the Cupertino Icons font to your application.
+#   # Use with the CupertinoIcons class for iOS style icons.
+#   cupertino_icons: ^1.0.0
+
+#   intl: ^0.16.1
+#   fl_chart: ^0.12.2
+#   flutter_svg: ^0.19.0
+#   floating_bottom_navigation_bar: ^1.4.1
+#   responsive_flutter: ^0.0.4
+#   # flutter_screenutil: ^4.0.2+5
+#   # device_preview: ^0.5.5
+#   device_preview: ^1.2.0
+
+# dev_dependencies:
+#   flutter_test:
+#     sdk: flutter
+
+# # For information on the generic Dart part of this file, see the
+# # following page: https://dart.dev/tools/pub/pubspec
+
+# # The following section is specific to Flutter.
+# flutter:
+#   # The following line ensures that the Material Icons font is
+#   # included with your application, so that you can use the icons in
+#   # the material Icons class.
+#   uses-material-design: true
+
+#   # To add assets to your application, add an assets section, like this:
+#   assets:
+#     - assets/
+#     - assets/imag/
+
+#   # An image asset can refer to one or more resolution-specific "variants", see
+#   # https://flutter.dev/assets-and-images/#resolution-aware.
+
+#   # For details regarding adding assets from package dependencies, see
+#   # https://flutter.dev/assets-and-images/#from-packages
+
+#   # To add custom fonts to your application, add a fonts section here,
+#   # in this "flutter" section. Each entry in this list should have a
+#   # "family" key with the font family name, and a "fonts" key with a
+#   # list giving the asset and other descriptors for the font. For
+#   # example:
+#   # fonts:
+#   #   - family: Schyler
+#   #     fonts:
+#   #       - asset: fonts/Schyler-Regular.ttf
+#   #       - asset: fonts/Schyler-Italic.ttf
+#   #         style: italic
+#   #   - family: Trajan Pro
+#   #     fonts:
+#   #       - asset: fonts/TrajanPro.ttf
+#   #       - asset: fonts/TrajanPro_Bold.ttf
+#   #         weight: 700
+#   #
+#   # For details regarding fonts from package dependencies,
+#   # see https://flutter.dev/custom-fonts/#from-packages
+
 name: personal_expenses_2
 description: A new Flutter project.
 
-# The following line prevents the package from being accidentally published to
-# pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none"
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.4.3 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  
-  intl: ^0.16.1
-  fl_chart: ^0.12.2
-  flutter_svg: ^0.19.0
-  floating_bottom_navigation_bar: ^1.4.1
-  responsive_flutter: ^0.0.4
-  # flutter_screenutil: ^4.0.2+5
-  device_preview: ^0.5.5
+  intl: ^0.19.0
+  fl_chart: ^0.69.0
+  uuid: ^3.0.6
+  # flutter_svg: ^2.0.10+1
+  # floating_bottom_navigation_bar: ^1.4.1
+  # # responsive_flutter: ^0.0.4
+  # device_preview: ^0.7.0 # Updated version
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-# For information on the generic Dart part of this file, see the
-# following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
-  # The following line ensures that the Material Icons font is
-  # included with your application, so that you can use the icons in
-  # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
     - assets/imag/
-    
-
-  # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
-
-  # For details regarding adding assets from package dependencies, see
-  # https://flutter.dev/assets-and-images/#from-packages
-
-  # To add custom fonts to your application, add a fonts section here,
-  # in this "flutter" section. Each entry in this list should have a
-  # "family" key with the font family name, and a "fonts" key with a
-  # list giving the asset and other descriptors for the font. For
-  # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
-  #
-  # For details regarding fonts from package dependencies,
-  # see https://flutter.dev/custom-fonts/#from-packages
