@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:personal_expenses_2/constants.dart';
 import 'package:personal_expenses_2/home_navigator.dart';
 import 'package:personal_expenses_2/screens/home.dart';
-import 'constants.dart';
 
 void main() {
   runApp(MyApp());
@@ -19,14 +18,17 @@ class MyApp extends StatelessWidget {
       //
       theme: ThemeData(
         primaryColor: kPrimaryColor,
-        accentColor: kAccentColor,
+        hintColor: kAccentColor,
         primarySwatch: kSwatchColor,
       ),
       debugShowCheckedModeBanner: false,
       title: 'Personal Expenses',
       home: HomeNavigator(),
       routes: {
-        Home.routeName: (ctx) => Home(),
+        Home.routeName: (ctx) => Home(
+              updateIncome: () {},
+              transactions: [], // Provide default or real data here
+            ),
         HomeNavigator.routeName: (ctx) => HomeNavigator(),
       },
     );
