@@ -1,8 +1,13 @@
 class Transaction {
-  final String id;
+  String id;
   final String title;
   final double amount;
   final DateTime date;
   final Map category;
-  Transaction({this.id, this.title, this.amount, this.date, this.category});
+  Transaction(
+      {required this.id,
+      required this.title,
+      required this.amount,
+      required this.date,
+      required this.category});
 }
